Seperate functions from main page file
